email_worker/pyproject.toml
[tool.poetry]
name = "email_worker"
version = "0.1.0"
description = "receiving and processing e-mai"
authors = ["VlNagibin"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
aio-pika = "^9.5.7"
pydantic-settings = "^2.6.1"
imaplib2 = "^3.6"
cryptography = "^45.0.4"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.2"
mypy = "^1.15.0"
isort = "^6.0.1"
pre-commit = "^4.1.0"

[tool.poetry.group.tests.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-subtests = "^0.11.0"
requests = "^2.32.3"
httpx = "^0.28.1"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79
force_grid_wrap = 0
use_parentheses = true
force_sort_within_sections = false
multi_line_output = 3
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = 3.12
follow_imports = "skip"
explicit-package-bases = true
namespace-packages = true

[tool.pytest.ini_options]
# asyncio_mode=auto
norecursedirs = "env/* venv/*"
addopts = [
    "-vv",
    "-p",
    "no:cacheprovider",
    "--strict-markers",
]
testpaths = "tests/"
python_files = "test_*.py"
markers = [
    "test",
    "core"
]
